generating temp files for bytes and results.

+++ java/mlAutoVirt/src/org/duo/ml/autovirt/InputAppState.java
@@ -131,7 +131,7 @@ public class InputAppState
         this.speed = speed;
     }
 
-    public float getAngleIndex() {
+    public int getAngleIndex() {
         return angleIndex;
     }
